+++ docs/source/changelog.md
@@ -1,9 +1,9 @@
 # Change log
 ## PreRelease
 ### Added 
+- Added charactersafe & existing scenario name checks (in InputChecker Class), resulting in a disabled create button
 ### Changed
 - Moved callback for dataset_name_invalid to inputchecker.py to avoid duplicate code in datamanagerderive/importmodal
-### Fixed
 
 ## 0.3.21
 _12-02-2026_